REFACTOR - being refactor to util fns
